[DEL] deleted unused project store file
